rename: line buffer
