+++ backend/modules/claude/line-buffer.ts
